pyproject.toml
[project]
name = "stock-prediction-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "tensorflow>=2.20.0",
    "uvicorn>=0.34.0",
]
